feat: stats() 映射引擎内存统计，补两条用例，M2 收口 (#6)
* feat: stats() 映射引擎内存统计，补 catch 内中断与跨线程调用用例，M2 收口

- kmpjs_stats 增加 JS_ComputeMemoryUsage 的已用字节、限额、对象 / 字符串 / atom / 函数计数；JsEngineStats 对应扩展
- 用例：中断不可被脚本的 catch 拦截；引擎在一个线程创建、另一个线程使用时栈检查仍正确（深递归报 RangeError）
- 决策记录补上 256 KB 栈限额约合 270 层 JS 递归的实测；shim ABI 升到 5


* fix: stats 的 memoryLimit 直接报配置值，避免 32 位目标把无限制误报成 4 GiB


---------

+++ library/src/androidMain/kotlin/wang/harlon/quickjs/NativeEngine.android.kt
@@ -52,7 +52,7 @@ internal actual class NativeEngine actual constructor(config: JsEngineConfig, in
     actual fun evalModule(source: String, name: String, flags: Int): RawValue =
         result(NativeBridge.nativeEvalModule(ptr, Wtf8.encode(source), Wtf8.encode(name), flags))
 
-    actual fun stats(): IntArray = NativeBridge.nativeStats(ptr) ?: throw JsException("native call failed")
+    actual fun stats(): LongArray = NativeBridge.nativeStats(ptr) ?: throw JsException("native call failed")
 
     actual fun dumpMemory(): RawValue = result(NativeBridge.nativeDumpMemory(ptr))
 }
